--- a/server/package.json
+++ b/server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "server",
-  "version": "0.4.9",
+  "version": "0.5.0",
   "description": "",
   "main": "index.js",
   "scripts": {