… README.md
… client/package.json
… client/src/App.min.css
… client/src/components/home.component.js
… client/src/components/login.component.js
… client/src/components/navbar.component.js
… client/src/components/register-oauth.component.js
… client/src/components/register.component.js
… client/src/components/req-oauth.component.js
… client/src/components/reset-password.component.js
… client/src/components/verify-account.component.js
… package.json
… server/index.js
… server/lib/callback.js
… server/lib/passport.js
… server/models/todo.model.js
… server/models/users.model.js
… server/routes/oauth.route.js
… server/routes/todo.route.js
… server/routes/users.route.js