**What's Changed in v0.5.0:**
1. Added `backup Codes` Feature (#126) (95547b6388beffc58740f113455ceec24c069828)
    * Added `Backup Codes` Method on `Login` Component (14aa26ac0aa1a1b2de423dd779357e6d3bfa1d2d)
    * Added Regenerate `Token` in `Client` (cb4fd72ad2be4b608e2af7a439a4d03bb4c0f5ba)
    * Added Preview `Backup Tokens` in `Client` (3a89ff34c2e7832a7edc732f72d1c21d3bcdf37f)
    * Added Generate `Backup Codes` in `Server` (5b25aab16033f126a522c58dba81e444b43f73ef)
    * Added Verify Backup `Codes` in `Server` (356b76f9b49941649170195d4ca0dab3ccfb289d)
    * Generate `Token` on Activating `MFA` (e08bea4abf172a66de3a2d6fd7ba2de7f9d80202)
    * Added `Backup Code` Method on `Login` Component (14aa26ac0aa1a1b2de423dd779357e6d3bfa1d2d)
2. Upgrade `Dependencies` to the Latest Version (aa11e8659b74989bb3aa0ecd20c3dbfae24310d3)
3. Improved `Performance` (40a8e91fba324fd36694a3f4cda93d118cae6a84)
4. Improved User Experiences (`UX`) (40a8e91fba324fd36694a3f4cda93d118cae6a84)
5. Improved `Select` Appearance (#122) (4268a7f38ae03547b64349cb6862cf4aeec8396c)
6. Resolved `Error` in Backup Codes `Modal` (d53fdedd23842f134c8f9249fd1676e27fda0fbf)
7. Improved Security: Limit `Request` with Rate Limiter (#123) (d5228102aca40677b91bb966e2c2e736a90c97c2)
8. Support for Longer `Credentials` (4ccb81598b1e7a269a64eb5274db8739167c5743)
9. Resolved Issues of:
    * No Messages are Returned After `Logout` and Activating `MFA` (#121) (a9f39314e84ab5f5d8bb06e6cd5d6f93358ab3c1)
    * Return `JSON` Message if `CSRF` Token or Rate Limit Reached (#124) (602228a0712b95394f35fff850c981112c7c0626)
    * Received `undefined` Email Route while Redirecting via `OAuth` (#125) (5e33e14d8fc38649bc0534d0c9cebe6675119d73)
    * `Issues` of Receiving `undefined` Messages in `Forgot Password` (c1fb3d4da9f29b50ad7d5117ebddcec41c91d521)
    * Fixed Minor Bug

Details About Changes in v0.5.0:
| Client | `build\static\js\2.6ca8d93c.chunk.js` | `build\static\js\main.2591d28a.chunk.js` | `build\static\css\main.030489ff.chunk.css` | `build\static\js\3.668acfa5.chunk.js` | `build\static\js\runtime-main.3cbd5898.js`
------ | --------------------------------------- | ------------------------------------------- | ---------------------------------------------- | -------------------------------------------------- | ---------------------------------------------- |
| v0.5.0 | 152.27 kB (+3.16 kB) | 28.34 kB (+428 B) | 4.54 kB (+37 B) | 1.56 kB | 1.16 kB |
| v0.4.9 | 149.11 kB (+159 B) | 27.92 kB (+115 B) | 4.5 kB (+1 B) | 1.56 kB | 1.16 kB |
| v0.4.8 | 148.95 kB (+42.58 kB) | 27.81 kB (+2.2 kB) | 4.5 kB (+224 B) | 1.56 kB (+2 B) | 1.16 kB (+2 B) |
| v0.4.7 | 148.29 kB (-1.25 kB) | 26.54 kB (+478 B) | 4.47 kB (+72 B) | 1.56 kB | 1.16 kB |
| v0.4.6 | 147.82 kB | 26.07 kB | 4.4 kB | 1.56 kB | 1.16 kB |
| v0.4.5 | 106.38 kB | 25.61 kB | 4.29 kB | 1.56 kB | 1.16 kB |

| Server | Size |
| ------- | ---- |
| v0.5.0 | 219.15 kB (+7.2 kB) |
| v0.4.9 | 211.95 kB (+95 B) |
| v0.4.8 | 211 kB (+21 kB) |
| v0.4.7 | 190 kB (+14 kB) |
| v0.4.6 | 176 kB |

For more information about changes in this release, please visit our [project roadmaps v0.5.0](https://github.com/stanleyowen/todo-application/projects/30)